allmias/requirements.txt
django
djangorestframework
django_resized
opencv-python
django_cleanup
numpy
django-cors-headers